Create a package 'rpc' for common functionalities between 'core' and 'cluster'
