+++ rpc/cluster/ClusterDatabase.java
@@ -19,11 +19,11 @@
  * under the License.
  */
 
-package com.vaticle.typedb.client.cluster;
+package com.vaticle.typedb.client.rpc.cluster;
 
 import com.vaticle.typedb.client.api.database.Database;
-import com.vaticle.typedb.client.core.TypeDBDatabaseImpl;
-import com.vaticle.typedb.client.core.TypeDBDatabaseManagerImpl;
+import com.vaticle.typedb.client.rpc.TypeDBDatabaseImpl;
+import com.vaticle.typedb.client.rpc.TypeDBDatabaseManagerImpl;
 import com.vaticle.typedb.protocol.ClusterDatabaseProto;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
