+++ rpc/cluster/ClusterSession.java
@@ -19,13 +19,13 @@
  * under the License.
  */
 
-package com.vaticle.typedb.client.cluster;
+package com.vaticle.typedb.client.rpc.cluster;
 
 import com.vaticle.typedb.client.api.TypeDBOptions;
 import com.vaticle.typedb.client.api.TypeDBSession;
 import com.vaticle.typedb.client.api.TypeDBTransaction;
 import com.vaticle.typedb.client.api.database.Database;
-import com.vaticle.typedb.client.core.TypeDBSessionImpl;
+import com.vaticle.typedb.client.rpc.TypeDBSessionImpl;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
